memory: add the memory spec and implement M1 (int8 arrival tapes)
New docs/spec/memory_optimize.md lays out M1-M5 with measurements, the
R0/contract rules, and acceptance. M1 implemented here.

M1: the dense arrival tape is (480,3,2,6,6) with max count 5, but was int32.
- generate_scenario now produces int8 (range-checked; OOD flood x20 = 100,
  capacity = 45) and data/io.py casts older int32 tapes on load.
- scenario_digest normalizes to int32 before hashing, so scenario_hash and
  every committed fixture are unchanged.
- crates/bus-sim-py accepts PyReadonlyArray5<i8> and widens to i32 when packing;
  the sparse store and kernel are unchanged.

Results (full seed, same-session pair):
- 600 scenarios: 249 MB -> 67 MB (tape 405 -> 101 KB each)
- peak RSS: Python 662 -> 479 MB, Rust 690 -> 508 MB (Rust 1236 -> 508 overall)
- speedup: Python 595.35 s vs Rust 161.61 s learn+val (3.68x), total 597.99 vs
  164.28 s (3.64x)
- byte-identical: 20-point validation curve, weights and optimizer match both
  across backends and int32->int8 on the Python backend
- verify OK, 149 passed; frozen native revision aedc4faa4c43

M2-M5 (sparse canonical representation, Rust-only ownership, store dedup,
lazy/mmap loading) are documented as follow-ups; they were left unstarted
because they overlap with M1's effort or risk the R3 contract.

+++ src/bus_rl/backend/native.py
@@ -73,7 +73,7 @@ def build_kernel(scenario: Scenario, *, conservation: bool = True):
     require_native()
     import bus_sim
 
-    arrivals = np.ascontiguousarray(scenario.arrival_tape, dtype=np.int32)
+    arrivals = np.ascontiguousarray(scenario.arrival_tape, dtype=np.int8)
     traffic = np.ascontiguousarray(scenario.traffic_tape, dtype=np.float32)
     kernel = bus_sim.Kernel(scenario_payload(scenario), arrivals, traffic)
     kernel.set_conservation_checks(conservation)
@@ -93,7 +93,7 @@ class NativeScenarioStore:
 
         self._store = bus_sim.ScenarioStore()
         for scenario in scenarios:
-            arrivals = np.ascontiguousarray(scenario.arrival_tape, dtype=np.int32)
+            arrivals = np.ascontiguousarray(scenario.arrival_tape, dtype=np.int8)
             traffic = np.ascontiguousarray(scenario.traffic_tape, dtype=np.float32)
             self._store.add(scenario_payload(scenario), arrivals, traffic)
